Move static functions to util
